event/cancel & event/delete
logic and functionality for cancel has been implemented. still need to setup relational database for events/users to handle cancellation properly

delete endpoint has been added and is functional so when an event is expired or the date for the event has passed we can remove it from the database.

event/register now adds an event to the database. again, still need to setup relational database for users and events. comment on line 30 of eventController explaining

#17