Working registration
Users can now register for events.
Events are handled properly when a user registers for an event.
If the event is already in the table, we just register the user.
If the event is not already in the table, we add it to the table, and register the user to the event.
#17
There is still a bug caused by a foreign key constraint when a user tries to register for an event that they are already registered for.
Users can now unregister (cancel) from events as well.